requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
SQLAlchemy
alembic
python-dotenv
python-jose
passlib[bcrypt]